refactor: view 관련 객체 코드 수정
