+++ src/views/OutputView.js
@@ -1,21 +1,20 @@
 import PROGRESS_MESSAGE from '../constants/messages/progressMessage.js';
 import RESULT_MESSAGE from '../constants/messages/resultMessage.js';
-import DELIMITER from '../constants/delimiters/delimter.js';
+import DELIMITER from '../constants/delimiter.js';
 
 const OutputView = {
   printStartGame() {
-    console.log(RESULT_MESSAGE.result_start);
+    console.log(RESULT_MESSAGE.START_PREFIX);
   },
 
   /**
    * 게임 라운드마다의 누적 결과를 출력합니다.
    * @param { Map } gameResult
    */
-
   printResult(gameResult) {
     gameResult.forEach((round) => {
       Object.entries(round).forEach(([key, value]) =>
-        console.log(key + DELIMITER.COLON + DELIMITER.HYPHEN.repeat(value)),
+        console.log(key + DELIMITER.COLON + DELIMITER.SCORE_MARK.repeat(value)),
       );
       this.divideLine();
     });
@@ -26,10 +25,10 @@ const OutputView = {
    * @param { Map } gameResult
    * @param { Number } count
    */
-
   printWinner(finalWinnerArr) {
     console.log(
-      PROGRESS_MESSAGE.FINAL_WINNER_PREFIX + finalWinnerArr.join(DELIMITER.COMMA + DELIMITER.SPACE),
+      PROGRESS_MESSAGE.FINAL_WINNER_PREFIX +
+        finalWinnerArr.join(DELIMITER.CAR_NAME_SEPARATOR + DELIMITER.SPACE),
     );
   },
